feat: add `ClientDescription` return type
This creates a friendlier interface for framework integration.

Instead of returning a `SourceDescription` from a client provider, you
can return a `ClientDescription` that works at a higher level. When you
do this, Saus handles the extraction of isomorphic renderer code from
the render module, and you get to focus on the hydration code and any
necessary imports.