chore: remove `cacheKey` method
This approach won't work, because the client doesn't know what the cacheKey is. We would need a lookup table, or else we'd have to fetch the state every time even when it can be reused from a page with the same cache key but a different URL.

So instead, a page is now allowed to derive state from the search query part of its URL.

This limitation won't exist in SSR environments, but statically generated pages will always have this limitation.