feat: add `saus.bundle.moduleMap` option
This option lets you store client modules outside the SSR bundle.

This commit also prepares the infrastructure for defining custom loaders for assets and JS modules. See the "assetLoader.ts" and "moduleLoader.ts" files for clues on how that will work.

Part of that work includes special treatment for binary assets. For example, binary assets are now properly encoded with Base64 if included in the SSR bundle. In addition, the `RenderedPage` object type that represents the result of a `renderPage` call now includes an `assets` property, whose values may be raw data buffers or HttpRedirect objects. Many of the helper functions for creating an HTTP server around your SSR bundle have also changed to accomodate this new handling of binary assets.

Lastly, HTML "post" plugins are now able to simply add an asset ID to the `state.assets` collection, and it will be loaded by the associated page automatically. So the responsibility of injecting a <link> element does not fall on the HTML plugin anymore.

As a result, the <link> and <script> tags injected by Saus will not be visible to HTML post plugins, and so those tags cannot be manipulated.