refactor: remove CompileCache worker
I have decided that cleaning up the cache directory in response to
unused files is not worth the effort (at least currently). The only
time this really matters is when a project file is deleted while
Saus is not running, which should rarely happen.

We *could* validate the cache if a new Git branch is detected,
but it is not a priority for now.